backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
# psycopg2-binary removed for SQLite migration
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
joblib>=1.4.0
pydantic>=2.5.0
